zcore: switch zcore_response to enum class

+++ exch/zcore/rpc_ext.cpp
@@ -1962,7 +1962,7 @@ BOOL rpc_ext_push_response(const RPC_RESPONSE *presponse,
 
 	if (!ext_push.init(nullptr, 0, EXT_FLAG_WCOUNT | EXT_FLAG_ZCORE))
 		return FALSE;
-	QRF(ext_push.p_uint8(zcore_response::SUCCESS));
+	QRF(ext_push.p_uint8(static_cast<uint8_t>(zcore_response::SUCCESS)));
 	if (EXT_ERR_SUCCESS != presponse->result) {
 		if (ext_push.p_uint32(4) != EXT_ERR_SUCCESS ||
 		    ext_push.p_uint32(presponse->result) != EXT_ERR_SUCCESS)
